# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Beauty Platform API
  version: v1
  description: API contract for Beauty Platform.
servers:
  - url: /api/v1
security:
  - CookieAuth: []
  - BearerAuth: []
...
paths:
...
  /crm/appointments:
    post:
      tags: [crm]
      summary: Create appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrmAppointmentCreateRequest'
      responses:
        '200':
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrmAppointmentResponse'
  /crm/appointments/{id}:
    patch:
      tags: [crm]
      summary: Update appointment
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrmAppointmentUpdateRequest'
      responses:
        '200':
          description: Appointment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrmAppointmentResponse'
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: session
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CrmAppointmentResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/CrmAppointment'
    CrmAppointment:
      type: object
      required: [id, specialistId, locationId, clientId, startAt, endAt, status]
      properties:
        id:
          type: integer
        specialistId:
          type: integer
        locationId:
          type: integer
        clientId:
          type: integer
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        status:
          type: string
        source:
          type: string
        clientName:
          type: string
        serviceNames:
          type: array
          items:
            type: string
        serviceIds:
          type: array
          items:
            type: integer
        priceTotal:
          type: string
        durationMin:
          type: integer
    CrmAppointmentCreateRequest:
      type: object
      required: [staffId, locationId, startAt, endAt]
      properties:
        staffId:
          type: integer
        locationId:
          type: integer
        clientId:
          type: integer
          nullable: true
        clientName:
          type: string
          nullable: true
        clientPhone:
          type: string
          nullable: true
        clientEmail:
          type: string
          nullable: true
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        status:
          type: string
          nullable: true
        priceTotal:
          type: string
          nullable: true
        serviceId:
          type: integer
          nullable: true
    CrmAppointmentUpdateRequest:
      type: object
      properties:
        staffId:
          type: integer
        locationId:
          type: integer
        clientId:
          type: integer
          nullable: true
        clientName:
          type: string
          nullable: true
        clientPhone:
          type: string
          nullable: true
        clientEmail:
          type: string
          nullable: true
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        status:
          type: string
          nullable: true
        priceTotal:
          type: string
          nullable: true
        serviceId:
          type: integer
          nullable: true
...